refactor: headQuarter -> headquarter로 리팩토링
