+++ src/main/java/uos/uos25/headquarter/repository/HeadQuarterRepository.java
@@ -1,7 +1,7 @@
-package uos.uos25.headQuarter.repository;
+package uos.uos25.headquarter.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import uos.uos25.headQuarter.entity.HeadQuarter;
+import uos.uos25.headquarter.entity.HeadQuarter;
 
 public interface HeadQuarterRepository extends JpaRepository<HeadQuarter, Long> {}
