+++ src/main/java/uos/uos25/headquarter/service/HeadquarterService.java
@@ -1,12 +1,12 @@
-package uos.uos25.headQuarter.service;
+package uos.uos25.headquarter.service;
 
 import java.util.List;
 
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
-import uos.uos25.headQuarter.entity.HeadQuarter;
-import uos.uos25.headQuarter.repository.HeadQuarterRepository;
+import uos.uos25.headquarter.entity.HeadQuarter;
+import uos.uos25.headquarter.repository.HeadQuarterRepository;
 
 @Service
 @RequiredArgsConstructor
