+++ src/main/java/uos/uos25/order/dto/response/DisposalCreateResponseDTO.java
@@ -2,7 +2,7 @@ package uos.uos25.order.dto.response;
 
 import lombok.Builder;
 import lombok.Data;
-import uos.uos25.headQuarter.dto.ProductInfo;
+import uos.uos25.headquarter.dto.ProductInfo;
 import uos.uos25.order.entity.Disposal;
 import uos.uos25.shop.dto.ShopInfo;
