+++ src/main/java/uos/uos25/order/dto/response/PurchaseCancelResponseDTO.java
@@ -1,4 +1,4 @@
-package uos.uos25.purchase.dto.response;
+package uos.uos25.order.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
