+++ src/main/java/uos/uos25/statistics/dto/response/StatisticsReceiptResponseDTO.java
@@ -1,7 +1,7 @@
 package uos.uos25.statistics.dto.response;
 
 import lombok.Data;
-import uos.uos25.receipt.dto.response.ReceiptGetResponseDTO;
+import uos.uos25.user.receipt.dto.response.ReceiptGetResponseDTO;
 
 @Data
 public class StatisticsReceiptResponseDTO {
