+++ src/main/java/uos/uos25/statistics/dto/response/StatisticsReturnstResponseDTO.java
@@ -1,7 +1,7 @@
 package uos.uos25.statistics.dto.response;
 
 import lombok.Data;
-import uos.uos25.returns.dto.response.ReturnsGetResponseDTO;
+import uos.uos25.order.dto.response.ReturnsGetResponseDTO;
 
 @Data
 public class StatisticsReturnstResponseDTO {
